Move files around (and update package.json) to make it v2.0.0
