+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "push-wrapper",
-  "version": "1.2.0",
+  "version": "2.0.0",
   "main": "push.js",
   "description": "A wrapper for the Ableton Push to give easy to use APIs for controlling/being controlled by the Push",
   "keywords": [
